pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ibis-birdbrain"
version = "0.0.9"
authors = [
  { name="Cody", email="cody@dkdc.dev" },
]
description = "data + AI"
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
  'ipython',
  'python-dotenv',
  'toml',
  'Pillow',
  'typer[all]',
  'requests',
  'html2text',
  'openai',
  'marvin[ddg]~=1.3.0',
  'ibis-framework[duckdb,polars,deltalake]',
  'xlsx2csv',
  'XlsxWriter',
  'plotly',
  'streamlit',
]

[project.urls]
"Homepage" = "https://github.com/ibis-project/ibis-birdbrain"
"Bug Tracker" = "https://github.com/ibis-project/ibis-birdbrain/issues"

[project.scripts]
birdbrain = "ibis_birdbrain.app:app"

